requirements.txt
# FastAPI 和 Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy>=2.0.27
psycopg2-binary==2.9.9

# HTTP 请求和解析
requests==2.31.0
httpx[http2]==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3

# 数据处理
pydantic==2.5.0
python-dotenv==1.0.0

# 文件处理
aiofiles==23.2.1

# Selenium 相关
selenium==4.15.2
webdriver-manager==4.0.1

# Playwright 相关
playwright==1.40.0

# 其他工具
python-dateutil==2.8.2
watchdog==3.0.0
croniter==2.0.1
pytz==2023.3
psutil==5.9.5
